mv dataset viewer rendering to module with the dataset loader
